Removed extra folders and moved helpers back to their folder. Fixed typo that was causing requests not to work
